renaming abstract_model.py to model.py
