--- opa/storage/abstract_model.py
@@ -1,20 +1,0 @@
-from typing import Dict
-from abc import ABC, abstractmethod
-
-
-class Entity(ABC):
-    """
-    Simple value object to capture information
-    """
-    pass
-
-
-class HbaseEntity(Entity):
-
-    @abstractmethod
-    def value(self) -> Dict:
-        pass
-
-    @abstractmethod
-    def id(self) -> str:
-        pass
